refactor : mainPage model sale,product,style로 분리, 서비스 로직 수정
